+++ js/react/jsPlugins.js
@@ -55,7 +55,7 @@
     /**
      * Define the [app] Object. This allows for the framework standard API call
      * [app.addPlugin(name, object)] to be used and allows for code from a React
-     * app to call [app.refreshJsPlugins()] to run the plugins.
+     * app to call [app.refreshPlugins()] to run the plugins.
      * 
      * Additionally several common functions and properties such as [showErrorAlert()]
      * are copied from the standard DataFormsJS framework so they can be used here
@@ -133,9 +133,9 @@
     
         /**
          * Call this to run all loaded plugins. Example usage:
-         *     <JsonData onViewUpdated={window.app.refreshJsPlugins}>
+         *     <JsonData onViewUpdated={window.app.refreshPlugins}>
          */
-        refreshJsPlugins: function() {
+        refreshPlugins: function() {
             // Only update if not already running
             if (isUpdating) {
                 return;