+++ js/web-components/url-router.js
@@ -30,6 +30,11 @@ import {
     usingWebComponentsPolyfill
 } from './utils.js';
 
+const appEvents = {
+    routeChanged: 'app:routeChanged',
+    error: 'app:error',
+};
+
 /**
  * Shadow DOM for Custom Elements
  */
@@ -51,8 +56,10 @@ function downloadTemplate(router, view, route, urlParams) {
     // Validate
     const url = route.src;
     if (url === null || url === '') {
-        showError(view, `Missing <template> or [src] attribute for route [${route.path}].`);
-        router.dispatchEvent(new Event('contentLoaded'));
+        const error = `Missing <template> or [src] attribute for route <${route.tagName.toLowerCase()} path="${route.path}">.`;
+        showError(view, error);
+        router.dispatchEvent(new CustomEvent(appEvents.error, { bubbles: true, detail: error }));
+        router.dispatchEvent(new Event(appEvents.routeChanged, { bubbles: true }));
         return;
     }
 
@@ -79,8 +86,9 @@ function downloadTemplate(router, view, route, urlParams) {
         setView(router, view, html, urlParams);
     })
     .catch(error => {
-        showError(view, error);
-        router.dispatchEvent(new Event('contentLoaded'));
+        const text = render`Error with <${route.tagName.toLowerCase()} path="${route.path}"> - Error Downloading Template: [${url}], Error: ${error}`;
+        showError(view, text);
+        router.dispatchEvent(new CustomEvent(appEvents.error, { bubbles: true, detail: text }));
     });
 }
 
@@ -146,7 +154,7 @@ function setView(router, view, html, urlParams) {
     }
 
     // Custom Event
-    router.dispatchEvent(new Event('contentLoaded'));
+    router.dispatchEvent(new Event(appEvents.routeChanged, { bubbles: true }));
 }
 
 /**
@@ -201,7 +209,7 @@ class UrlRouter extends HTMLElement {
             return;
         }
 
-		// Get URL Path
+        // Get URL Path
         let path = window.location.pathname;
         if (path === '') {
             path = '/';
@@ -317,8 +325,8 @@ class UrlRouter extends HTMLElement {
         this.style.color = 'white';
         this.style.fontSize = '1.5em';
         this.textContent = message;
-        this.dispatchEvent(new Event('error'));
-        this.dispatchEvent(new Event('contentLoaded'));
+        this.dispatchEvent(new CustomEvent(appEvents.error, { bubbles: true, detail: message }));
+        this.dispatchEvent(new Event(appEvents.routeChanged, { bubbles: true }));
         console.error(message);
     }
 }